% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_box.R
\name{fire_box}
\alias{fire_box}
\title{Fire Emission Boxplot}
\usage{
fire_box(Fires, i)
}
\arguments{
\item{Fires}{Dataframe containing DOI and USFS fire and emissions data.}

\item{i}{value of 1 "DOI" or 2 "USFS" for data source}
}
\description{
This function reads in a dataframe of fire and emissions data from DOI and USFS and creates box plots by state.
}
\examples{
fire_box(Fires = x, i = 1)
}
\keyword{Emissions}
\keyword{Fire}
\keyword{Histogram}
