% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_hist.R
\name{fire_hist}
\alias{fire_hist}
\title{Fire Emission Histograms}
\usage{
fire_hist(fires)
}
\arguments{
\item{fires}{Dataframe containing DOI and USFS fire and emissions data.}
}
\description{
This function reads in a datafram of fire and emissions data from DOI and USFS and creates histograms by state.
}
\examples{
fire_hist(fires = x)
}
\keyword{Emissions}
\keyword{Fire}
\keyword{Histogram}
